+++ R/token.R
@@ -2,7 +2,10 @@
 #'
 #' Create a token
 #'
-#' @param n Number of bytes.
+#' @param n Integer /// Optional. \cr
+#'          Number of bytes. \cr
+#'          Defaults to `16L`.
+#'
 #' @examples
 #' token_create()
 #' token_create(n = 32L)